tests: separate into own files
